requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==65.0.1
django-anymail==12.0
django-rest-knox==5.0.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
